validations

+++ app/Http/Controllers/RegistrationController.php
@@ -6,6 +6,9 @@ use Illuminate\Http\Request;
 
 use App\Http\Requests;
 
+use App\Meeting;
+use App\User;
+
 class RegistrationController extends Controller
 {
     /**
@@ -16,33 +19,41 @@ class RegistrationController extends Controller
      */
     public function store(Request $request)
     {
+      $this->validate($request, [
+        'meeting_id' => 'required',
+        'user_id' => 'required'
+      ]);
+
       $meeting_id = $request->input('meeting_id');
       $user_id = $request->input('user_id');
 
-      $meeting = [
-        'title' => 'Title',
-        'description' => 'Description',
-        'time' => 'time',
-        'user_id' => 'User Id',
-        'view_meeting' => [
-          'href' => 'api/v1/meeting/1',
-          'method' => 'GET'
+      $meeting = Meeting::findOrFail($meeting_id);
+      $user = User::findOrFail($user_id);
+
+      $message = [
+        'msg' => 'User is already registered for Meeting',
+        'meeting' => $meeting,
+        'user' => $user,
+        'unregister' => [
+          'href' => 'api/v1/meeting/registration/' . $meeting->id,
+          'method' => 'DELETE',
         ]
       ];
+      if ($meeting->users()->where('users.id', $user->id)->first()) {
+        return response()->json($message, 404);
+      }
 
-      $user = [
-        'name' => 'Name'
-      ];
+      $user->meetings()->attach($meeting);
 
       $response = [
         'msg' => 'User registered for Meeting',
         'meeting' => $meeting,
         'user' => $user,
         'unregister' => [
-          'href' => 'api/v1/meeting/registration/1',
+          'href' => 'api/v1/meeting/registration/' . $meeting->id,
           'method' => 'DELETE',
         ]
-      ]
+      ];
 
       return response()->json($response, 200);
     }
@@ -55,31 +66,19 @@ class RegistrationController extends Controller
      */
     public function destroy($id)
     {
-      $meeting = [
-        'title' => 'Title',
-        'description' => 'Description',
-        'time' => 'time',
-        'user_id' => 'User Id',
-        'view_meeting' => [
-          'href' => 'api/v1/meeting/1',
-          'method' => 'GET'
-        ]
-      ];
-
-      $user = [
-        'name' => 'Name'
-      ];
+      $meeting = Meeting::findOrFail($id);
+      $meeting->users()->detach();
 
       $response = [
         'msg' => 'User unregistered for Meeting',
         'meeting' => $meeting,
-        'user' => $user,
-        'unregister' => [
+        'user' => 'tbd',
+        'register' => [
           'href' => 'api/v1/meeting/registration',
           'method' => 'POST',
           'params' => 'user_id, meeting_id'
         ]
-      ]
+      ];
 
       return response()->json($response, 200);
     }
